# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DocuForge API
  version: 1.0.0
  description: OpenAPI specification for DocuForge core services (auth, users, documents, conversions, security, admin)
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /documents/{id}/download:
    get:
      summary: Download document file
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File stream
...